Move data folder inside dataprep directory.
This is much easier than relying on the previous idea of using a constants module. From now on, all data to pre processed just has to be placed in the correct folder, and the program will just ask before conversion what the starting format is (and will empty all downstream directories).